requirements.txt
requests==2.31.0
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.1
python-dateutil>=2.8.0